pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "crewai>=0.118.0",
    "crewai-tools>=0.0.1",
    "fastapi>=0.115.12",
    "jinja2>=3.1.6",
    "openai>=1.86.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.3",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "pydantic[email]>=2.0.0",
    "fastapi-limiter>=0.1.6",
    "redis>=5.0.0",
]
